requested changes

+++ server/src/main/java/game/ai/StupidBot.java
@@ -50,9 +50,9 @@ public class StupidBot extends Bot {
         }
 
         if (!turnClockwise) {
-            alpha = alpha + turningRate > Math.PI ? (float) -Math.PI : alpha + turningRate;
+            alpha = alpha + turningRate > Math.PI ? -Math.PI : alpha + turningRate;
         } else {
-            alpha = alpha - turningRate < -Math.PI ? (float) Math.PI : alpha - turningRate;
+            alpha = alpha - turningRate < -Math.PI ? Math.PI : alpha - turningRate;
         }
 
         if (counter > changeDirectionAtCounter) {
